+++ src/test/debugger/extension/attachQuickPick/provider.unit.test.ts
@@ -6,17 +6,11 @@
 import * as assert from 'assert';
 import { expect } from 'chai';
 import { anything, instance, mock, verify, when } from 'ts-mockito';
-import { Disposable } from 'vscode';
-import { ApplicationShell } from '../../../../client/common/application/applicationShell';
-import { CommandManager } from '../../../../client/common/application/commandManager';
-import { IApplicationShell, ICommandManager } from '../../../../client/common/application/types';
-import { Commands } from '../../../../client/common/constants';
 import { PlatformService } from '../../../../client/common/platform/platformService';
 import { IPlatformService } from '../../../../client/common/platform/types';
 import { ProcessService } from '../../../../client/common/process/proc';
 import { ProcessServiceFactory } from '../../../../client/common/process/processFactory';
 import { IProcessService, IProcessServiceFactory } from '../../../../client/common/process/types';
-import { IDisposableRegistry } from '../../../../client/common/types';
 import { OSType } from '../../../../client/common/utils/platform';
 import { AttachProcessProvider } from '../../../../client/debugger/extension/attachQuickPick/provider';
 import { PsProcessParser } from '../../../../client/debugger/extension/attachQuickPick/psProcessParser';
@@ -24,13 +18,9 @@ import { IAttachItem } from '../../../../client/debugger/extension/attachQuickPi
 
 // tslint:disable-next-line: max-func-body-length
 suite('Attach to process - process provider', () => {
-    let applicationShell: IApplicationShell;
-    let commandManager: ICommandManager;
     let platformService: IPlatformService;
     let processService: IProcessService;
     let processServiceFactory: IProcessServiceFactory;
-    let disposableRegistry: IDisposableRegistry;
-    let disposable: Disposable;
 
     let provider: AttachProcessProvider;
 
@@ -43,33 +33,13 @@ suite('Attach to process - process provider', () => {
 `;
 
     setup(() => {
-        disposable = mock(Disposable);
-        applicationShell = mock(ApplicationShell);
-
-        commandManager = mock(CommandManager);
-        when(commandManager.registerCommand(Commands.PickLocalProcess, anything(), anything())).thenReturn(instance(disposable));
-
         platformService = mock(PlatformService);
-
         processService = mock(ProcessService);
         when(processService.exec(anything(), anything(), anything())).thenResolve({ stdout: psOutput });
         processServiceFactory = mock(ProcessServiceFactory);
         when(processServiceFactory.create()).thenResolve(instance(processService));
 
-        disposableRegistry = [];
-
-        provider = new AttachProcessProvider(instance(applicationShell), instance(commandManager), disposableRegistry, instance(platformService), instance(processServiceFactory));
-    });
-
-    teardown(() => {
-        (disposableRegistry as Disposable[]).forEach(d => d.dispose());
-    });
-
-    test('registerCommands() should register the show quick pick command', () => {
-        provider.registerCommands();
-
-        verify(commandManager.registerCommand(Commands.PickLocalProcess, anything(), anything())).once();
-        assert.equal((disposableRegistry as Disposable[]).length, 1);
+        provider = new AttachProcessProvider(instance(platformService), instance(processServiceFactory));
     });
 
     test('The Linux process list command should be called if the platform is Linux', async () => {
